존재하지않는 게시물입니다. 게시물 수정시 수정 날짜도 갱신 findByTitle

+++ src/main/java/com/mysite/sbb/article/controller/ArticleController.java
@@ -8,6 +8,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -48,6 +49,8 @@ public class ArticleController {
             article.setBody(body); //불러온 데이터 수정
         }
 
+        article.setUpdateDate(LocalDateTime.now());
+
         articleRepository.save(article); //수정된 데이터 db에 저장
 
         return article;
@@ -56,10 +59,20 @@ public class ArticleController {
     @RequestMapping("/doDelete")
     @ResponseBody
     public String doDelete(long id) {
+        if(articleRepository.existsById(id) == false) {
+            return "%d번 게시물은 이미 삭제되었거나 존재하지 않습니다.".formatted(id);
+        }
+
         articleRepository.deleteById(id); // 삭제
         return "%d번 게시물이 삭제되었습니다".formatted(id);
 
     }
 
+    @RequestMapping("findByTitle")
+    @ResponseBody
+    public List<Article> findByTitle(String title) {
+        List<Article> articles = articleRepository.findByTitle(title);
+        return articles;
+    }
 
 }
